feat: metadata image
